Refactored buffers, added scene rendering support

+++ Flatter/include/rendering/Mesh.h
@@ -14,7 +14,7 @@ class Mesh {
   Mesh(Rendering::Vulkan::IndexedVertexBufferRef vertexBuffer);
   Mesh(const Rendering::Vulkan::ContextRef& context,
        const std::vector<glm::vec3>& vertices,
-       const std::vector<unsigned int>& indices);
+       const std::vector<uint16_t>& indices);
 
   const Rendering::Vulkan::IndexedVertexBufferRef& getIndexedVertexBuffer()
       const {
